fix post body reactivity to post changes
